+++ NEWS.md
@@ -16,7 +16,7 @@ Version 0.5.1
 
 ### Other changes
 
-* Updated bundled KImageFormats plugins to version from KDE Frameworks 6.10.
+* Updated bundled KImageFormats plugins to version from KDE Frameworks 6.19.
 
 
 Version 0.5.0